+++ src/minizinc/model.py
@@ -72,6 +72,7 @@ class Model:
     _enum_map: Dict[str, Enum]
     _includes: List[Path]
     _lock: threading.Lock
+    _checker: bool = False
 
     def __init__(self, files: Optional[Union[ParPath, List[ParPath]]] = None):
         self._data = {}
@@ -171,10 +172,12 @@ class Model:
                 )
                 with self._lock:
                     self._includes.append(file)
-        elif file.suffix != ".mzn":
+        elif file.suffix not in [".mzn", ".mzc"]:
             raise NameError("Unknown file suffix %s", file.suffix)
         else:
             with self._lock:
+                if ".mzc" in file.suffixes:
+                    self._checker = True
                 self._includes.append(file)
 
     def add_string(self, code: str) -> None:
@@ -192,42 +195,3 @@ class Model:
         copy._code_fragments = self._code_fragments[:]
         copy._data = dict.copy(self._data)
         return copy
-
-
-class Checker:
-    """An interface to a MiniZinc solution checker
-
-    Attributes:
-        checker (Path): The file path to the solution checker file.
-
-    """
-
-    checker: Path
-
-    def __init__(self, path: Path):
-        self.checker = path
-        assert self.checker.exists()
-
-    def check(self, solution) -> str:
-        """Check the Solution object using Checker
-
-        This method runs the Checker model on the solution provided. The
-        output from the checker model is returned as a string.
-
-        Args:
-            solution (Union[Solution, Mapping[str, Any]]): The solution given to
-                the checker
-        """
-        from minizinc import Solver
-        from minizinc.CLI import CLIInstance
-
-        solver = Solver.lookup("gecode")
-        instance = CLIInstance(solver)
-        instance.add_file(self.checker)
-        for i in instance.input.keys():
-            if isinstance(solution, Mapping):
-                instance[i] = solution[i]
-            else:
-                instance[i] = getattr(solution, i)
-
-        return str(instance.solve())